backend/requirements.txt
asgiref==3.5.0
Django==4.0.10
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-gis==1.0
drf-yasg==1.21.6
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
psycopg2==2.9.6
pytz==2023.3
sqlparse==0.4.4
pytest==7.4.0
pytest-django==4.5.2
python-dotenv==1.0.0
